Move config module to ipatests.pytest_plugins.integration.config
https://pagure.io/freeipa/issue/6798
