+++ ipatests/ipa-test-config
@@ -27,7 +27,8 @@ import argparse
 import json
 
 from ipalib.constants import FQDN
-from ipatests.test_integration import config, env_config
+from ipatests.test_integration import env_config
+from ipatests.pytest_plugins.integration import config
 
 
 def main(argv):
