+++ ipatests/pytest_plugins/integration/__init__.py
@@ -30,6 +30,7 @@ from pytest_multihost import make_multihost_fixture
 
 from ipapython import ipautil
 from ipapython.ipa_log_manager import log_mgr
+from .config import Config
 
 log = log_mgr.get_logger(__name__)
 
@@ -154,7 +155,6 @@ def mh(request, class_integration_logs):
     """
     # TODO: cleanup modules
     from ipatests.test_integration import tasks
-    from ipatests.test_integration.config import Config
     from ipatests.test_integration.env_config import get_global_config
 
     cls = request.cls
